Automatically add LifecycleModule to all injectors created by Governator
